+++ nextstep-mvc/src/main/java/nextstep/mvc/tobe/adapter/HandlerAdapter.java
@@ -1,6 +1,6 @@
 package nextstep.mvc.tobe.adapter;
 
-import nextstep.mvc.tobe.ModelAndView;
+import nextstep.mvc.tobe.view.ModelAndView;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
