+++ nextstep-mvc/src/main/java/nextstep/mvc/tobe/adapter/LegacyHandlerAdapter.java
@@ -1,8 +1,8 @@
 package nextstep.mvc.tobe.adapter;
 
 import nextstep.mvc.asis.Controller;
-import nextstep.mvc.tobe.JspView;
-import nextstep.mvc.tobe.ModelAndView;
+import nextstep.mvc.tobe.view.JspView;
+import nextstep.mvc.tobe.view.ModelAndView;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
